[engraving] added score layout class
